# perf(processor/post): improve processing speed when `config.post_asset_folder` is enabled