--- a/lib/plugins/processor/post.ts
+++ b/lib/plugins/processor/post.ts
@@ -1,13 +1,14 @@
 import { toDate, timezone, isExcludedFile, isTmpFile, isHiddenFile, isMatch } from './common';
 import Promise from 'bluebird';
 import { parse as yfm } from 'hexo-front-matter';
-import { extname, join, posix } from 'path';
+import { extname, join, posix, sep } from 'path';
 import { stat, listDir } from 'hexo-fs';
 import { slugize, Pattern, Permalink } from 'hexo-util';
 import { magenta } from 'picocolors';
 import type { _File } from '../../box';
 import type Hexo from '../../hexo';
 import type { Stats } from 'fs';
+import { PostSchema } from '../../types';
 
 const postDir = '_posts/';
 const draftDir = '_drafts/';
@@ -270,26 +271,35 @@ function processAsset(ctx: Hexo, file: _File) {
   const id = file.source.substring(ctx.base_dir.length);
   const doc = PostAsset.findById(id);
 
-  if (file.type === 'delete') {
+  if (file.type === 'delete' || Post.length === 0) {
     if (doc) {
       return doc.remove();
     }
-
     return;
   }
 
-  if (Post.length > 0) {
-    const assetDir = id.slice(0, id.lastIndexOf('/'));
-    const post = Post.findOne(p => p.asset_dir.endsWith(posix.join(assetDir, '/')));
+  const savePostAsset = (post: PostSchema) => {
+    return PostAsset.save({
+      _id: id,
+      slug: file.source.substring(post.asset_dir.length),
+      post: post._id,
+      modified: file.type !== 'skip',
+      renderable: file.params.renderable
+    });
+  };
+
+  if (doc) {
+    // `doc.post` is `Post.id`.
+    const post = Post.findById(doc.post);
     if (post != null && (post.published || ctx._showDrafts())) {
-      return PostAsset.save({
-        _id: id,
-        slug: file.source.substring(post.asset_dir.length),
-        post: post._id,
-        modified: file.type !== 'skip',
-        renderable: file.params.renderable
-      });
+      return savePostAsset(post);
     }
+  }
+
+  const assetDir = id.slice(0, id.lastIndexOf(sep));
+  const post = Post.findOne(p => p.asset_dir.endsWith(posix.join(assetDir, '/')));
+  if (post != null && (post.published || ctx._showDrafts())) {
+    return savePostAsset(post);
   }
 
   if (doc) {