@yoshinorin 2024-04-20
@hexojs/core 
Dear team. This PR is ready for review.
Thanks :)