@D-Sketon 2025-10-05
The Windows sep is a backslash. Assuming the ID is `source/_posts/test-2/1.png`, after slicing the assetDir becomes `source/_posts/test-2/1.pn`. This clearly indicates an issue.